requirements.txt
Flask==3.0.0
Pillow==10.1.0
google-generativeai==0.3.1